+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 
 ### Fixed
 ### Changed
+
+- move from facsimile to Ctl (#32)
+
 ### Deprecated
 ### Removed
 ### Security