+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 
 ## Unreleased
+
+
+## 1.7.0
 ### Added
 - tests for django 2.1
 - fix #33: easy namespace inheritance through related models